03_rpiVersion/splendid/requirements.txt
Adafruit-Blinka==8.33.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-cap1188==1.3.12
adafruit-circuitpython-connectionmanager==1.0.0
adafruit-circuitpython-requests==3.0.0
adafruit-circuitpython-typing==1.10.2
adafruit-extended-bus==1.0.2
Adafruit-PlatformDetect==3.61.0
Adafruit-PureIO==1.1.11
getmac==0.9.4
joblib==1.3.2
numpy==1.26.4
paho-mqtt==1.6.1
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
sysv-ipc==1.1.0
threadpoolctl==3.3.0
typing_extensions==4.10.0
